Falcor 5.1 (#296)

+++ Source/Falcor/Scene/Importers/USDImporter/Utils.h
@@ -1,5 +1,5 @@
 /***************************************************************************
- # Copyright (c) 2015-21, NVIDIA CORPORATION. All rights reserved.
+ # Copyright (c) 2015-22, NVIDIA CORPORATION. All rights reserved.
  #
  # Redistribution and use in source and binary forms, with or without
  # modification, are permitted provided that the following conditions
@@ -141,7 +141,7 @@ namespace Falcor
         }
         else
         {
-            logError("Unsupported primvar interpolation mode {}.", freq);
+            logError("Unsupported primvar interpolation mode {}.", (uint32_t)freq);
             return 0;
         }
     }
